+++ extensions/cursor-directory/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Cursor Directory Changelog
 
-## [Refactors and Improvements] - 2024-09-10
+## [Refactors and Improvements] - 2024-09-15
 
 - Fetch prompts from API instead of github repo
 - Fix display of prompt details
+- Allow configure if show detailed view in prompts list via preferences
+- Allow configure default prompts list to show via preferences
+- Add count of how many times a prompt has been copied on cursor.directory for popular prompts
 
 ## [Initial Version] - 2024-09-10